+++ src/index.ts
@@ -1,1 +1,2 @@
 export { ProxyBalancer } from "./proxyBalancer";
+export { getProxyURL } from "./utils/proxy";